+++ spec/fixtures/dummy/config/routes.rb
@@ -46,6 +46,8 @@ Rails.application.routes.draw do
     post "nested_params/nested", to: "nested_params#nested"
     post "nested_params/empty_block", to: "nested_params#empty_block"
     post "nested_params/non_hash_block", to: "nested_params#non_hash_block"
+    get "rescued_resources/:id", to: "rescued_resources#show"
+    get "rescued_resources_with_view", to: "rescued_resources_with_view#index"
     post "inputs/upload", to: "inputs#upload"
     resources :inputs, only: %i[show create]
   end